requirements.txt
setuptools==76.0.0
datasets==3.2.0
huggingface-hub==0.30.2
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
tokenizers==0.21.1
torch==2.6.0
transformers==4.51.3
accelerate==1.4.0
catboost==1.2.7
requests==2.32.3
python-dateutil==2.9.0.post0
matplotlib==3.9.4
python-dotenv==1.0.1
tqdm==4.67.1
numpy==1.26.4
semgrep==1.114.0
openai==1.70.0
cityhash==0.4.8
ollama==0.4.7
optuna==4.3.0
boto3==1.38.4
boto3-stubs[essential]==1.38.4
botocore==1.38.8
tensorboard==2.19.0